@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 
 /**
  * This file is part of me-cms-instagram.
@@ -15,9 +16,11 @@
 namespace MeCmsInstagram\Test\TestCase\Controller;
 
 use Cake\Cache\Cache;
+use Cake\Controller\Controller;
+use Cake\Event\EventInterface;
 use Cake\ORM\Entity;
-use MeCmsInstagram\Controller\Component\InstagramComponent;
 use MeCms\TestSuite\ControllerTestCase;
+use MeCmsInstagram\Controller\Component\InstagramComponent;
 
 /**
  * InstagramControllerTest class
@@ -28,7 +31,7 @@ class InstagramControllerTest extends ControllerTestCase
      * Internal method to get a mock instance of `InstagramComponent`
      * @return \MeCmsInstagram\Controller\Component\InstagramComponent|\PHPUnit\Framework\MockObject\MockObject
      */
-    protected function getInstagramComponentMock()
+    protected function getInstagramComponentMock(): object
     {
         $methods = [
             'getMediaResponse' => 'media.json',
@@ -47,11 +50,11 @@ class InstagramControllerTest extends ControllerTestCase
 
     /**
      * Adds additional event spies to the controller/view event manager
-     * @param \Cake\Event\Event $event A dispatcher event
+     * @param \Cake\Event\EventInterface $event A dispatcher event
      * @param \Cake\Controller\Controller|null $controller Controller instance
      * @return void
      */
-    public function controllerSpy($event, $controller = null)
+    public function controllerSpy(EventInterface $event, ?Controller $controller = null): void
     {
         parent::controllerSpy($event, $controller);
 
@@ -66,6 +69,7 @@ class InstagramControllerTest extends ControllerTestCase
      */
     public function testBeforeRender()
     {
+        $this->disableErrorHandlerMiddleware();
         $this->get(['_name' => 'instagramPhotos']);
         $this->assertInstanceof(Entity::class, $this->viewVariable('user'));
 
@@ -80,7 +84,7 @@ class InstagramControllerTest extends ControllerTestCase
     {
         $this->get(['_name' => 'instagramPhotos']);
         $this->assertResponseOkAndNotEmpty();
-        $this->assertTemplate('Instagram' . DS . 'index.ctp');
+        $this->assertTemplate('Instagram' . DS . 'index.php');
         $this->assertContainsOnlyInstancesOf(Entity::class, $this->viewVariable('photos'));
 
         $nextIdFromView = $this->viewVariable('nextId');
@@ -88,7 +92,7 @@ class InstagramControllerTest extends ControllerTestCase
 
         //Sets the cache name
         $cache = sprintf('index_limit_%s', getConfigOrFail('default.photos'));
-        list($photosFromCache, $nextIdFromCache) = array_values(Cache::read($cache, 'instagram'));
+        [$photosFromCache, $nextIdFromCache] = array_values(Cache::read($cache, 'instagram'));
 
         $this->assertEquals($this->viewVariable('photos'), $photosFromCache);
         $this->assertEquals($nextIdFromView, $nextIdFromCache);
@@ -96,14 +100,14 @@ class InstagramControllerTest extends ControllerTestCase
         //GET request. Now with the `nextId`
         $this->get(['_name' => 'instagramPhotosId', $nextIdFromView]);
         $this->assertResponseOkAndNotEmpty();
-        $this->assertTemplate('Instagram' . DS . 'index.ctp');
+        $this->assertTemplate('Instagram' . DS . 'index.php');
 
         //GET request. Now it's an ajax request, with the `nextId`
         $this->configRequest(['headers' => ['X-Requested-With' => 'XMLHttpRequest']]);
         $this->get(['_name' => 'instagramPhotosId', $nextIdFromView]);
         $this->assertResponseOkAndNotEmpty();
-        $this->assertTemplate('Instagram' . DS . 'index.ctp');
-        $this->assertLayout('Layout' . DS . 'ajax' . DS . 'ajax.ctp');
+        $this->assertTemplate('Instagram' . DS . 'index.php');
+        $this->assertLayout('layout' . DS . 'ajax' . DS . 'ajax.php');
     }
 
     /**
@@ -115,7 +119,7 @@ class InstagramControllerTest extends ControllerTestCase
 
         $this->get(['_name' => 'instagramPhoto', $id]);
         $this->assertResponseOkAndNotEmpty();
-        $this->assertTemplate('Instagram' . DS . 'view.ctp');
+        $this->assertTemplate('Instagram' . DS . 'view.php');
         $this->assertInstanceof(Entity::class, $this->viewVariable('photo'));
 
         $photoFromCache = Cache::read(sprintf('media_%s', md5($id)), 'instagram');
